# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple withdrawal and inquiry service",
        "title": "Simple Withdraw API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:6005",
    "basePath": "/api",
    "paths": {
        "/api/withdraw/{userId}": {
            "get": {
                "description": "Get withdrawal records by user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Withdrawal"
                ],
                "summary": "Get withdrawal history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Withdrawal"
                            }
                        }
                    }
                }
            }
        },
        "/tools/signature": {
            "post": {
                "description": "Generate SHA-256 signature based on method, relative path, body, and timestamp.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tools"
                ],
                "summary": "Generate Signature",
                "parameters": [
                    {
                        "description": "Signature Request Payload (param is request body for protected service)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SignatureRequestExample"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SignatureResponseDto"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "description": "Get list of all users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.User"
                            }
                        }
                    }
                }
            }
        },
        "/user/create": {
            "post": {
                "description": "Create a new user and generate initial balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create new user",
                "parameters": [
                    {
                        "description": "Create User Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserRequestDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/user/{userId}": {
            "get": {
                "description": "Get a specific user by user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.User"
                        }
                    }
                }
            }
        },
        "/withdraw/all": {
            "get": {
                "description": "Get all withdrawal transactions in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Withdrawal"
                ],
                "summary": "Get all withdrawal records",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Withdrawal"
                            }
                        }
                    }
                }
            }
        },
        "/withdraw/create": {
            "post": {
                "description": "Create a withdrawal transaction for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Withdrawal"
                ],
                "summary": "Create a withdrawal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request timestamp",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Request signature",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Withdrawal Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.WriteTransactionRequestDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "withdrawal successful",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.Withdrawal": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateUserRequestDto": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.SignatureRequestExample": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object",
                    "properties": {
                        "amount": {
                            "type": "integer",
                            "example": 10000
                        },
                        "userId": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                },
                "method": {
                    "type": "string",
                    "example": "POST"
                },
                "relativePath": {
                    "type": "string",
                    "example": "/api/withdraw/create"
                },
                "secretKey": {
                    "type": "string",
                    "example": "dev-secret"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T12:00:00Z"
                }
            }
        },
        "dto.SignatureResponseDto": {
            "type": "object",
            "properties": {
                "signature": {
                    "type": "string",
                    "example": "f7321c4c3c99..."
                }
            }
        },
        "dto.WriteTransactionRequestDto": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "userId": {
                    "type": "integer"
                }
            }
        }
    }
}